+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 <sub>TBD</sub>
 
 ### Updates
-- Parse and encode selector placeholders [#11](https://github.com/adamgian/minify-selectors/issues/11)
+- Parse and encode minify-selectors specific prefixed selectors [#11](https://github.com/adamgian/minify-selectors/issues/11)
 - Add support for JS Location interface proprties: `window.location.hash`, 'window.location.href' and `window.location`
 
 ### Fixes